refactor: 도메인 중심으로 패키지 구조 변경
