+++ src/main/java/roomescape/reservation/ReservationController.java
@@ -10,7 +10,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import roomescape.service.ReservationService;
 
 @Controller
 @RequestMapping("/reservations")
