+++ src/main/java/roomescape/time/ReservationTimeController.java
@@ -10,7 +10,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import roomescape.service.ReservationTimeService;
 
 @Controller
 @RequestMapping("/times")
